refactor: useFetchWalletAddress hooks
